third_party/WebKit/LayoutTests/http/tests/htmlimports/import-upgrade-iframe.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
var proto = Object.create(HTMLElement.prototype);
proto.createdCallback = function () {
    this.isCreated = true;
};

document.registerElement('x-foo', { prototype: proto });
</script>
</head>
<body>
<!-- This contains slow import but it shouldn't block anything in this document -->
<iframe src="resources/linking-slow-import.html"></iframe>
<!-- Let the parser wait for a while to ensure the iframe content be parsed -->
<script src="resources/do-nothing-js.cgi"></script>
<x-foo id=target></x-foo>
<script>
shouldBeTrue('target.isCreated');
</script>
</body>
</html>
