third_party/WebKit/LayoutTests/http/tests/htmlimports/import-upgrade-iframe-expected.txt

PASS target.isCreated is true
PASS successfullyParsed is true

TEST COMPLETE
